Document OWASP ZAP review process
Sprint 7 incorporates Docker-based OWASP ZAP testing. Flexion is running it on their end to ensure that there are no errors. This could be gated as a part of the CD pipeline, but that would only work if there were _no_ errors and, in fact, there are several errors that we consider acceptable. We may modify this process down the line (e.g., configure OWASP ZAP to ignore those errors), but for now, the process is for us to run OWASP ZAP in the testing process, and confirm that there are no errors that we have not previously said are acceptable (in the code review process, on a PR).

Former-commit-id: 9e84d6b91cc0b635ede8f9dc1c839a06c0950e31