Add a "merge to master" issue template
This adds a new template for GitHub issues, one that’s intended for proposing to move code from `staging` into `master`. It's a checklist of everything that needs to be provided in order to approve that pull request.

This would be better as a pull-request template, but GitHub doesn't support multiple pull-request templates like they do for issue templates, so we’ll start like this. Presumably GitHub will add that feature before we use this functionality many times.

Former-commit-id: 6fca056e0f8c7340d92b9f915e573da6955fa6cf