Change the link to open a new issue using the template
I think GitHub has changed how to open a new issue using a template. When this page was written, a template could be selected from the new issue page. Now they have an interstitial, which the prior link was bypassing. Here, instead, we link directly to a new issue, pre-populated with the template.

Former-commit-id: 4da4cb2d6b2eab44bf3837fb6415bda995f255bc
Former-commit-id: 45516bd8d8a4f732bc2122b33e0e18199f3979c9