Work can now be found in staging and master
Previously, it was necessary to specify that that `staging` holds all work, and `master` would show no activity. Now that we've merged work from `staging` to `master`, and intend to do so routinely, that's no longer the case. But now the absence of any badges for `master` is weird (there's no build status, etc.), so I've replaced the old disclaimer with a new disclaimer.