Cache keys from more than one user pool.
This fixes an error seen in the IRS environment:

```
TypeError: Expected "jwk" to be an Object
    at t (/var/task/cognito-authorizer.js:172:109)
    at i (/var/task/cognito-authorizer.js:524:512)
    at Runtime.exports.handler (/var/task/cognito-authorizer.js:524:969)
    at Runtime.handleOnce (/var/runtime/Runtime.js:66:25)
```

The issue is that the lambda caches the keys retrieved from the user
pool endpoint, and since the IRS is using two user pools simultaneously,
the second request fails as it is not signed by the first request’s user
pool keys.

The fix is to allow the lambda to cache keys from more than one user
pool, and look up the appropriate pool’s authorized keys (or fetch them)
when authorizing.


Former-commit-id: 8e74b11e19bcd52087b96aa84285a2e3fa924b7c
Former-commit-id: aab182890129e0237034e74f7ee0dce2887266ba